created favorite cards
